Split connection_matrix
